backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
joblib==1.3.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9  # For PostgreSQL support
gunicorn==21.2.0  # For production deployment 